Move selectAvailableSimulator to Simulator
